notif [nfc]: Split code into files "receive" and "display"
